Address review: use uint for length clamp and tighten test lengths
Convert the Unicode length prefix to uint before the bounds check so that
a value like 0xffffffff cannot become a negative int on 32-bit platforms
and panic in make. Also set the crafted test's layer/mask and layer info
lengths to their actual values (0x44/0x40) instead of the oversized 0xff,
keeping the case focused on the truncated luni payload.

+++ util.go
@@ -86,13 +86,15 @@ func readUnicodeString(b []byte) string {
 	if len(b) < 4 {
 		return ""
 	}
-	ln := int(readUint32(b, 0))
+	ln := uint(readUint32(b, 0))
 	if ln == 0 {
 		return ""
 	}
 	// ln counts UTF-16 code units, each 2 bytes, following the 4-byte length.
-	// Never read (or allocate) beyond what the buffer actually holds.
-	if avail := (len(b) - 4) >> 1; ln > avail {
+	// Never read (or allocate) beyond what the buffer actually holds. Using an
+	// unsigned type keeps the comparison correct on 32-bit platforms, where a
+	// length such as 0xffffffff would otherwise become a negative int.
+	if avail := uint((len(b) - 4) >> 1); ln > avail {
 		ln = avail
 	}
 	buf := make([]uint16, ln)
